api/requirements.txt

django>=1.8,<1.9
django-extensions>=1.5,<1.6
django-allauth>=0.23.0
djangorestframework>=3.2,<3.3
django-rest-swagger>=0.3,<0.4
django-rest-auth>=0.5.0
django-cors-headers>=1.1,<1.2
Pillow>=2.9,<2.10
gunicorn==19.3.0
mysqlclient>=1.3,<1.4
pdfminer3k>=1.3,<1.4
nltk>=3.1,<3.2